Updated Crc64Nvme to have a JS and VM implementation due to the int size difference between platforms
